Decompile InitializeStartup, UpdateStartup and ExitStartup and name their callees
The startup game mode (US 0x0801F264/0x0801F36C/0x0801F524) shows the
publisher slogan and legal notice screens with alpha-blend cross-fades, then
pushes MainMenu. Name the BG, blend and text helpers it calls
(ClearBgTilemap_candidate, LoadBgGraphic_candidate, SetAlphaBlendTargets,
SelectTextFont_candidate, DrawTextLines_candidate and others), label its three
BG control constants and graphic resources, and make GetLanguage return u32
and SetAlphaBlendCoefficients take u16 arguments.

+++ src/text/get_language.c
@@ -1,7 +1,7 @@
 #include "types.h"
 #include "text.h"
 
-u8 GetLanguage(void)
+u32 GetLanguage(void)
 {
     return gCurrentLanguage;
 }
